+++ dallinger/prolific.py
@@ -274,7 +274,7 @@ class ProlificService:
 
     def draft_study(
         self,
-        completion_code: str,
+        completion_codes: List[str],
         completion_option: str,
         description: str,
         eligibility_requirements: List[dict],
@@ -318,7 +318,7 @@ class ProlificService:
 
         # We can now create the draft study.
         payload = {
-            "completion_code": completion_code,
+            "completion_codes": completion_codes,
             "completion_option": completion_option,
             "description": description,
             "eligibility_requirements": eligibility_requirements,
@@ -345,7 +345,7 @@ class ProlificService:
 
     def create_study(
         self,
-        completion_code: str,
+        completion_codes: List[dict],
         completion_option: str,
         description: str,
         eligibility_requirements: List[dict],  # can be empty, but not None